requirements.txt
werkzeug==2.0.3
SQLALchemy==1.4.31
alembic==1.7.6
Flask==2.0.3
Flask-Login==0.5.0
Flask-SQLALchemy==2.5.1
flask-session==0.4.0
flask-principal==0.4.0
flask-migrate==3.1.0

click==8.0.4
xlsxwriter==3.0.3
openpyxl==3.0.9
xlrd==1.2.0
#openpyxl==2.6.4

xmltodict==0.12.0
PyYaml==6.0
Pillow==9.0.1
cachelib==0.6.0

huey==2.4.3
celery==5.2.3
pyaes==1.6.1

# MySQL-python; mysql8 pymysql client need cryptography
#mysql-connector==2.2.9
pymysql==1.0.2

requests==2.27.1
xpinyin==0.7.6

pystrich==0.8
qrcode==7.3.1
lxml==4.8.0
wheel==0.37.1
geopy==2.2.0
pycrypto==2.6.1

cryptography==2.1.4

# sudo apt-get install libmysqld-dev
# vi ~/.pip/pip.conf
#[global]
#trusted-host=pypi.douban.com
#index-url = http://pypi.douban.com/simple
